fix(dogfood): repair multimodal and value trust boundaries
