+++ pyproject.toml
@@ -65,7 +65,7 @@ full = [
     "uvicorn>=0.51.0",
 ]
 test = [
-    "pytest>=8,<9",
+    "pytest>=9.0.3,<10",
     "tiktoken>=0.9,<0.14",
 ]
 benchmark = [
